create separate container for boardname
